(vimrc) Nerdtree, Solarized Removed and Airline Update
Removed Nerdtree, as it was causing too many issues in combination with
kitty. Furthermore, I removed the Solarized theme, as I updated my kitty
config to use the Panda theme. Obviously, this was clashing with each
other. Finally, I updated the Airline part of the config to use the
theme luna instead of solarized.